% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_parse_solid.R
\name{rdf_parse_solid}
\alias{rdf_parse_solid}
\title{Parse RDF dataset from Solid server}
\usage{
rdf_parse_solid(docs, ..., rdf = NULL, client)
}
\arguments{
\item{docs}{One or more URLs to fetch from Solid server(s)}

\item{...}{not used}

\item{rdf}{optional existing rdflib rdf store}

\item{client}{client object, as returned by solid_client_register_dyn}
}
\value{
parsed triples in the form of an rdf object containing the redland world and model objects
}
\description{
Parse RDF dataset from Solid server
}
