% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{req_oauth_solid_dpop}
\alias{req_oauth_solid_dpop}
\title{OAuth authentication with authorization code}
\usage{
req_oauth_solid_dpop(req, client)
}
\arguments{
\item{req}{httr2 req object}

\item{client}{client object, as returned by solid_client_register_dyn}
}
\value{
httr2 req object with authorization headers
}
\description{
Uses oauth_flow_auth_code to generate an access token,
used to authenticate the request with oauth_client_req_auth_dpop. The token
is automatically cached to minimise the number of times login/consent is performed.
}
\examples{
\dontrun{
url <- "https://MY_PRIVATE_URL"
request(url) \%>\%
 req_oauth_solid_dpop(client) \%>\%
 req_perform() \%>\%
 resp_body_string()
}
}
