% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solid_client_register_clientid.R
\name{solid_client_register_clientid}
\alias{solid_client_register_clientid}
\title{Define Solid client using client id document}
\usage{
solid_client_register_clientid(IDP, client_id)
}
\arguments{
\item{IDP}{Identity provider}

\item{client_id}{URL to Client ID document, as described in https://solid.github.io/solid-oidc/primer/#authorization-code-pkce-flow-step-7}
}
\description{
Define Solid client using client id document
}
\seealso{
\link{solid_client_register_dyn}
}
