% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{solid_client_register_dyn}
\alias{solid_client_register_dyn}
\title{Register dynamic client}
\usage{
solid_client_register_dyn(IDP)
}
\arguments{
\item{IDP}{Identity provider}
}
\description{
Register dynamic client
}
